requirements.txt
opencv-python>=4.8.0
mediapipe
numpy
PyOpenGL
PyOpenGL_accelerate
pygame>=2.0.0
scikit-learn
pandas
google-generativeai
pillow
